Move proxy deploying to IdentityProxyDeploy
